+++ include/external/vcc/core/model/base_document_builder.hpp
@@ -9,8 +9,8 @@ namespace vcc
     {
         protected:
             // Error Message
-            GETSET(size_t, NumberOfCharactersBeforePosForErrorMessage, 10);
-            GETSET(size_t, NumberOfCharactersAfterPosForErrorMessage, 50);
+            GETSET(int64_t, NumberOfCharactersBeforePosForErrorMessage, 10);
+            GETSET(int64_t, NumberOfCharactersAfterPosForErrorMessage, 50);
             std::wstring GetErrorMessage(const std::wstring &str, const size_t &pos, const std::wstring &msg) const;
 
         protected:
